Closed Issue #1 and #8
#8:
Replaced ChecpPathUp() with MoveUpwards() and CheckForUpwards(), which is a co-routine that is running only when the player is moving.

Changes to is Moving can be done with setisMoving(bool) now and this will make sure to take care of the co-routine CheckForUpwards() and the animation controller for the idle animation as well.

#1:
Sphere rotation fixed and code refactored to have one method to call to rotate the sphere.