Continuation back in + Adjusted problems with going up.
Put continuation back in and fit it into the new structure of the movement code. Target selection and movement itself are not mostly separated.
Exceptions are:
When going up/down a new target is set within the movement after the gravity changes.

This caused problems in up and prevents continuation with down. I solved the problems in up by setting a new "start position".