Pause menu problem fixed
When confirming the continue or restart button with the controller this was done with the same button thats used for warp. Since Input events are evaluated before(!) Update, this ment that the InputDown for the Warp Button was True the whole frame. So the player would warp up right away.
To prevent this I changed the listeners and the UIController is now acting as interface for that and moves the evaluation into LateUpdate() which is called after Update().