Player behavior changes
Appear Spehre collision layer was changed so obstacles dont collide with it anymore. That way, the teleport animation is finished before dying when teleporting into a sphere.

SFX emmited by the player were set to 3D (default was 2D).

Statemachine now executes OnExitState on reset. This is a test to see if something breaks by doing that.