Fixed level transition camera bug
When the next level was loaded the camera would remain in the old level in default state until the level was loaded. As loading is very fast this caused a flicker.

Now after a level is won, the game wont resume (i.e. the TimeScale wont be reset) until the new level was loaded. This prevents the flickering.